+++ stable_baselines/ppo2/ppo2.py
@@ -20,13 +20,15 @@ class PPO2(BaseRLModel):
     :param policy: (ActorCriticPolicy) The policy model to use (MLP, CNN, LSTM, ...)
     :param env: (Gym environment or str) The environment to learn from (if registered in Gym, can be str)
     :param gamma: (float) Discount factor
-    :param n_steps: (int) The number of steps to run for each environment
+    :param n_steps: (int) The number of steps to run for each environment per update (i.e. batch size is n_steps * n_env where
+        n_env is number of environment copies running in parallel)
     :param ent_coef: (float) Entropy coefficient for the loss caculation
     :param learning_rate: (float or callable) The learning rate, it can be a function
     :param vf_coef: (float) Value function coefficient for the loss calculation
     :param max_grad_norm: (float) The maximum value for the gradient clipping
     :param lam: (float) Factor for trade-off of bias vs variance for Generalized Advantage Estimator
-    :param nminibatches: (int) Number of minibatches for the policies
+    :param nminibatches: (int) Number of training minibatches per update. For recurrent policies,
+        should be smaller or equal than number of environments run in parallel.
     :param noptepochs: (int) Number of epoch when optimizing the surrogate
     :param cliprange: (float or callable) Clipping parameter, it can be a function
     :param verbose: (int) the verbosity level: 0 none, 1 training information, 2 tensorflow debug